Logged in state is preserved despite page refresh, added loading page
Moved onAuthStateChanged event listener to componentDidMount(), there is a 1 second delay before the logged in state is saved to the React state. In this 1 second delay, the login/register button shows instead of the logout button.

To cover the 1 second delay before the logout button shows as it's supposed to, there is now a loading page that displays for that time.